+++ subscriptions_service/src/main/java/com/subOne/subscriptions_service/dto/analytic_subscription/SubscriptionLastDatePaymentDto.java
@@ -5,5 +5,5 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
-public record SubscriptionLastDatePaymentDto(Long subscriptionId, String paymentPeriod, BigDecimal amount, LocalDate datePaid) {
+public record SubscriptionLastDatePaymentDto(Integer subscriptionId, String paymentPeriod, BigDecimal amount, LocalDate datePaid) {
 }
